refactor(DAO-2135): move request-processing components into subfolder
Move ActiveRequestSection and RequestProcessingBlock (plus tests and
stories) from btc-vault root and components/ into request-processing/
subfolder. Add barrel index.ts.
