+++ src/app/btc-vault/components/request-processing/RequestProcessingBlock.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/nextjs'
 import { SectionContainer } from '@/app/communities/components/SectionContainer'
-import { RequestProcessingBlock } from './components/RequestProcessingBlock'
-import type { ActiveRequestDisplay } from './services/ui/types'
+import { RequestProcessingBlock } from './RequestProcessingBlock'
+import type { ActiveRequestDisplay } from '../../services/ui/types'
 
 const meta = {
   title: 'BTC Vault/RequestProcessingBlock',
