+++ src/app/btc-vault/components/request-processing/index.ts
@@ -1,0 +1,2 @@
+export { ActiveRequestSection } from './ActiveRequestSection'
+export { RequestProcessingBlock } from './RequestProcessingBlock'
